Function future::join10

pub fn join10<A, B, C, D, E, F, G, H, I, J, ERR>(fa: Future<A, ERR>, fb: Future<B, ERR>, fc: Future<C, ERR>, fd: Future<D, ERR>, fe: Future<E, ERR>, ff: Future<F, ERR>, fg: Future<G, ERR>, fh: Future<H, ERR>, fi: Future<I, ERR>, fj: Future<J, ERR>) -> Future<(A, B, C, D, E, F, G, H, I, J), ERR> where A: Debug + 'static, B: Debug + 'static, C: Debug + 'static, D: Debug + 'static, E: Debug + 'static, F: Debug + 'static, G: Debug + 'static, H: Debug + 'static, I: Debug + 'static, J: Debug + 'static, ERR: Debug + 'static