Struct future::Future

pub struct Future<A, E> where 'static: , E: 'static {
    // some fields omitted
}

A handle on the result of an asynchronous compution that allows for transformations and side effects.

Future<A, E> can be thought of as an asynchrounous Result with improved transformation and side-effect mechanisms. It is intended to be used to chain transformations and then eventually consumed. The two mechanisms for final consumption are blocking via future::await or future::await_safe, or with a side-effecting callback, added via resolve, resolve_success, or resolve_err.

Additionally, side-effects that don't consume the Future can be added via on_completion, on_success, and on_err.

Examples

use future;
use std::thread;

let (future, setter) = future::new::<i64, String>();
thread::spawn(move || {
    // some work here
    let result: Result<i64, String> = Ok(5);
    setter.set_result(result);
});

future
    .map(|i| i + 5)
    .map_err(|err| err + " more info")
    .on_err(|err| println!("Got an err: {:?}", err))
    .resolve(|answer| println!("Answer: {:?}", answer));

Methods

impl<A: 'static, E: 'static> Future<A, E>

fn is_resolved(&self) -> bool

Checks whether the result on the Future has been set

Examples

let (future, setter) = future::new::(); assert(future.is_resolved() == false); setter.set_result(Ok(0)); assert(future.is_resolved());

fn map<F, B>(self, f: F) -> Future<B, E> where F: FnOnce(A) -> B, F: 'static, B: 'static

Transform a successful value when the transformation cannot fail.

Examples

use future;
use future::Future;

let future_int: Future<i64, ()> = future::value(0);
let future_string: Future<String, ()> = future_int.map(|i| format!("{}", i));

fn map_err<F, E2>(self, f: F) -> Future<A, E2> where F: FnOnce(E) -> E2, F: 'static, E2: 'static

Transform an error value into another.

Examples

use future;
use future::Future;

struct MyError(String);

let f1: Future<(), String> = future::err(String::from("an error!"));
let f2: Future<(), MyError> = f1.map_err(|err_str| MyError(err_str));

fn handle<F>(self, f: F) -> Future<A, E> where F: FnOnce(E) -> A, F: 'static

Transform an error value into a success value.

Examples

use future;
use future::Future;

let future: Future<i64, String> = future::err(String::from("unknown"));
let handled_future = future.handle(|err| {
    if err == "unknown" {
        -1
    } else {
        0
    }
});
assert_eq!(-1, future::await(handled_future).unwrap());

fn and_then<F, B, E2>(self, f: F) -> Future<B, E> where F: FnOnce(A) -> Result<B, E2>, F: 'static, E2: Into<E>, E2: 'static, B: 'static

Transform a success value when the transformation might fail. Returns an Err if either the original computation or the transformation fail. The error type of the transformation must have an instance of Into so that the final result has the same error type.

Examples

use future;
use future::Future;
use std::num;

enum MyError {
    ParseError(num::ParseIntError)
}

impl From<num::ParseIntError> for MyError

let f1: Future<String, MyError> = future::value(String::from("4"));
let f2: Future<i64, MyError> = f1.and_then(|s| s.parse::<i64>());
assert_eq!(4, future::await(f2).unwrap());

fn rescue<F, E2>(self, f: F) -> Future<A, E> where F: FnOnce(E) -> Result<A, E2>, F: 'static, E2: Into<E>, E2: 'static

Like handle, except when the error transformation could fail.

fn transform<F, B, E2>(self, f: F) -> Future<B, E2> where F: FnOnce(Result<A, E>) -> Result<B, E2>, F: 'static, E2: 'static, B: 'static

The most general Future transformation; Transform the result of a Future, changing the success and error types if desired.

fn and_thenf<F, B, E2>(self, f: F) -> Future<B, E> where F: FnOnce(A) -> Future<B, E2>, F: 'static, E2: Into<E>, E2: 'static, B: 'static

Like and_then, except when the transformation returns another Future instead of a Result

fn bind<F, B, E2>(self, f: F) -> Future<B, E> where F: FnOnce(A) -> Future<B, E2>, F: 'static, E2: Into<E>, E2: 'static, B: 'static

Monadic bind; same as and_thenf

fn rescuef<F, E2>(self, f: F) -> Future<A, E> where F: FnOnce(E) -> Future<A, E2>, F: 'static, E2: Into<E>, E2: 'static

Like rescue, except when the transformation returns another Future instead of a Result

fn transformf<F, B, E2>(self, f: F) -> Future<B, E2> where F: FnOnce(Result<A, E>) -> Future<B, E2>, F: 'static, E2: 'static, B: 'static

Like transform, except when the transformation returns another Future instead of a Result

fn on_err<F>(self, f: F) -> Future<A, E> where F: FnOnce(&E) -> (), F: 'static

fn on_success<F>(self, f: F) -> Future<A, E> where F: FnOnce(&A) -> (), F: 'static

fn on_completion<F>(self, f: F) -> Future<A, E> where F: FnOnce(&Result<A, E>) -> (), F: 'static

fn resolve_success<F>(self, f: F) where F: FnOnce(A) -> (), F: 'static

Stores the side-effecting f to be run once the Future completes. f will only run if the Future resolves successfully; an error result will be dropped. This consumes the Future

fn resolve_err<F>(self, f: F) where F: FnOnce(E) -> (), F: 'static

Stores the side-effecting f to be run once the Future completes. f will only run if the Future resolves unsuccessfully; a successful result will be dropped. This consumes the Future

fn resolve<F>(self, f: F) where F: FnOnce(Result<A, E>) -> (), F: 'static

Stores the side-effecting f to be run once the Future completes. This consumes the Future, and is the most common method of consuming the final result of a Future computation.

impl<A, E, E2> Future<Future<A, E2>, E> where A: 'static, E: 'static, E2: Into<E> + 'static

fn flatten(self) -> Future<A, E>

Flatten a Future<Future<A, E2>, E> into a Future, where E: From<E2>